Tweak triggering condition
Depending on the kernel used, ConditionVirtualization=microsoft wasn’t always met, resulting in skipping the module on init. In that case, ConditionVirtualization=wsl seems to fix the issue. This patch  combines both possibilities into triggering condition (working like logical OR, e.g. unit is started if any of the conditions is met).